backend/pyproject.toml
[project]
name = "hoagiemeal"
version = "0.0.3"
description = "Meal by Hoagie"
readme = "README.md"
requires-python = ">=3.10, >=3.12"
license = {file = "../LICENSE"}

authors = [
    { name = "Hoagie Club", email = "hoagie@princeton.edu" },
]

maintainers = [
    { name = "Hoagie Club", email = "hoagie@princeton.edu" }
]

keywords = [
    "nutrient app",
    "nutrient tracker",
    "food diary",
    "meal planner",
    "diet tracker",
    "nutrient calculator",
    "healthy eating",
    "balanced diet",
    "food logging",
    "dietary management",
    "allergen tracker",
    "dietary restrictions",
    "dietary recommendations",
    "campus dining",
    "student wellness",
    "student nutrition",
    "calorie counting",
    "macro tracking",
    "meal logging",
    "goal setting",
    "progress tracking",
    "meal recommendations",
    "healthy habits",
    "progressive web app",
    "full stack web app",
]

classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: Unix"
]

dependencies = [
    "requests>=2.32.3",
    "ruff>=0.7.0",
    "django>=5.1",
    "django-cors-headers>=4.6.0",
    "dj-database-url>=2.3.0",
    "django-heroku>=0.3.1",
    "msgspec>=0.18.6",
    "python-dotenv>=1.0.1",
    "colorama>=0.4.6",
    "xmlschema>=3.4.2",
    "icalendar>=6.0.1",
    "lxml>=5.3.0",
    "beautifulsoup4>=4.12.3",
    "pgvector>=0.3.6",
    "djangorestframework>=3.15.2",
    "openai>=1.63.0",
    "orjson>=3.10.15",
    "tabulate>=0.9.0",
    "pytz>=2025.1",
    "pyjwt>=2.10.1",
]

# ===============================
# Project URLs
# ===============================
[project.urls]
Homepage = "https://meal.hoagie.io/"
Repository = "https://github.com/hoagieclub/meal/"

[tool.ruff]
# Exclude commonly ignored directories from linting
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

# General Ruff settings
line-length = 119 # Slight deviation from Google style since we are using Django.
indent-width = 4
target-version = "py312"

# -------------------------------
# Ruff Linting Rules
# -------------------------------
[tool.ruff.lint]

# Enable pydocstyle rules for Google-style docstrings
select = ["E4", "E7", "E9", "F", "B", "Q", "D"] # Flake8 rules, Bugbear, Quotes, etc.

ignore = [
    "D203",  # Ignore conflicts for one-liner docstring style
    "D213",  # Allow Google-style multi-line docstrings with different line breaks
]

# Define fixable and unfixable rules
fixable = ["ALL"]
unfixable = []

# -------------------------------
# Ruff Flake8-Quotes Settings
# -------------------------------
[tool.ruff.lint.flake8-quotes]
docstring-quotes = "double"  # Enforce double quotes for docstrings
inline-quotes = "double"     # Enforce double quotes for inline strings

# -------------------------------
# Ruff Formatting Settings
# -------------------------------
[tool.ruff.format]
quote-style = "double"                  # Use double quotes for auto-formatting
indent-style = "space"                  # Indent with spaces instead of tabs
skip-magic-trailing-comma = false       # Respect magic trailing commas
line-ending = "auto"                    # Automatically detect line endings

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["hoagiemeal"]

[tool.uv.workspace]
members = ["."]
